🚚 zb: Rename an internal file
